+++ packages/rsc/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ai-sdk/rsc
 
+## 3.0.0-beta.9
+
+### Patch Changes
+
+- Updated dependencies [1fe058b]
+  - ai@7.0.0-beta.9
+
+## 3.0.0-beta.8
+
+### Patch Changes
+
+- ai@7.0.0-beta.8
+
+## 3.0.0-beta.7
+
+### Patch Changes
+
+- Updated dependencies [210ed3d]
+  - ai@7.0.0-beta.7
+
 ## 3.0.0-beta.6
 
 ### Patch Changes